pyproject.toml
[project]
name = "ebay-to-shopify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ebayoauth==0.0.1",
    "httpx==0.28.1",
    "pandas==2.2.3",
    "python-dotenv>=1.1.0",
    "requests==2.32.3",
]
